refactor: 상품 엔티티 변경에 따른 관련 클래스 수정
